:arrow_up:Switch from Grunt to Webpack
Grunt is a fine task runner however Webpack is a bundler which can walk
dependency trees to figure out what needs to be included in the bundle.

Case in point, in this commit I told Webpack to include all the *.png files
but it only included the *.png which are actually used in the application.

This switch from Grunt to Webpack is a bit large because where Grunt ran
scripts and figured out how to compile the single JS file by reading what
scripts were included in the index.html Webpack walks the dependencies
based on what each JS file is importing. This means that not only did
I have to add the webpack.config but I also had to make sure that all the
JS files were importing their dependent modules.

To simplifly the importing this change refactored the JS files a bit by
seperating controller from routing definition and adding index.js files
to generate module bounderies, an down stream module (B) which depends
on an upstream module (A) only needs to import it's index.js file, internally
the dependent module (A) can be broken down into smaller internal modules
between multiple JS files while still presenting an easy to consume interface.

I tried to keep the changes mimimual but the change from a compile script to
a dependency bundler did require that I change quite a bit about how the
project was structured. So sorry about the size :crying_cat_face: