Fix nginx rewrites
Under the develop folder we have a compose file which defines a nginx config
this is to test running the application behind a proxy which routes based on
post fix paths (i.e. localhost:8080/docker-ui -> http://localhost:9000)

In the last commit where we added Webpack we have Webpack create the
index.html file, it was minifying said file so the base tag was being
written without quotations (<base href=/ >) which is different then
then nginx rewrite was expecting (<base href="/">) thus the base tag
was not getting rewritten when passing through the proxy.

The Rewrites are needed to insert the proxy postfix to the relative
urls, the urls to fetch bundle.css for example, so that if behind
a proxy which post fixes docker-ui to the URL thent he relative
URL is rewritten to read docker-ui/bundle.css without the application
needing to know about the post fix.

Without the rewrite when loading the application from behind a proxy
the browser was getting a 404 because it was looking for /bundles.css
vs. /docker-ui/bundles.css.

This change set fixes the rewrites to expect the minified base tag
and rewrite the base tag properly.