:bug:Fix _back_ button not disabeling
If you are on the first page the _back_ button should be disabled but
it was not. This bug was probably introduce when I updated the controllers
to use the controller as syntax (i.e. use Classes vs. functions). The root
cause was that a property of the controller was not initialized and
undefined is not less then 1 so the button was never disabled.

This change set fixes that by initializing the property to zero