Update proxy target in webpack config
The webpack config sets up the dev server with a proxy so that any requests
to /v2 will be proxied to the docker registry
(i.e. http://localhost:9000/v2/_catalog -> http://localhost:5000/v2/_catalog)

It was setup to read environment variables to tell where you have the registry
setup and on which port, it uses the same env var that the app uses.

However if you run the dev server on your host machine it is a pain to have
to remember to setup the env var so this change set updates the config file
to default to localhost:5000.

You can still use the env var to override the defaults.

This works with the setup of this project, we have a develop/docker-compose.yml
file which brings up a registry container on port 5000 and exposes that port
on the host so http://localhost:5000 will be that container. You can also
run the application via a container in that docker-compose file and it is
setup to set the required env var so win win their too.